A few fixes

+++ Pw.Elka.TIN/Server/Source Files/Components/ClientSession/Communicates/ServComADDRGETALL.cpp
@@ -13,25 +13,25 @@ ServComADDRGETALL::ServComADDRGETALL(vector<AddressModel>* addrVect)
 	unsigned char* myBuffer = communicateBuffer + 1;
 	communicateBuffer[0] = _SERVCOMADDRGETALL;			//communicate code
 	
-	myBuffer = storeShort(myBuffer, addrVect->size());
+	myBuffer = storeShort(myBuffer, (short)(addrVect->size()));
 	
-	for (unsigned int i = 0; i < addrVect->size(); ++i)
+	for (unsigned int i = 0; i < addrVect->size(); ++i) 
 	{
 		myBuffer = storeShort(myBuffer, 4);
 		myBuffer = storeInt(myBuffer, (*addrVect)[i].id);
 	}
-	myBuffer = storeShort(myBuffer, addrVect->size());
+	myBuffer = storeShort(myBuffer, (short)(addrVect->size()));
 
 	for (unsigned int i = 0; i < addrVect->size(); ++i)
 	{
-		myBuffer = storeShort(myBuffer, (*addrVect)[i].value.length());
+		myBuffer = storeShort(myBuffer, (short)((*addrVect)[i].value.length()));
 		myBuffer = storeString(myBuffer, (*addrVect)[i].value);
 	}
-	myBuffer = storeShort(myBuffer, addrVect->size());
+	myBuffer = storeShort(myBuffer, (short)(addrVect->size()));
 
 	for (unsigned int i = 0; i < addrVect->size(); ++i)
 	{
-		myBuffer = storeShort(myBuffer, (*addrVect)[i].name.length());
+		myBuffer = storeShort(myBuffer, (short)((*addrVect)[i].name.length()));
 		myBuffer = storeString(myBuffer, (*addrVect)[i].name);
 	}
