+++ Pw.Elka.TIN/Server/Source Files/Components/ClientSession/Communicates/ServComGRPGETONE.cpp
@@ -7,7 +7,7 @@ ServComGRPGETONE::ServComGRPGETONE(GroupModel *groupDB)
 	{
 		size = size + 2 + 4 + 2 + groupDB->addresses[i].value.size();
 	}
-
+	 
 	communicateBuffer = new unsigned char[size];
 	unsigned char* myBuffer = communicateBuffer + 1;
 	communicateBuffer[0] = _SERVCOMGRPGETONE;			//communicate code
@@ -15,21 +15,21 @@ ServComGRPGETONE::ServComGRPGETONE(GroupModel *groupDB)
 	myBuffer = storeShort(myBuffer, 4);
 	myBuffer = storeInt(myBuffer, groupDB->id);
 
-	myBuffer = storeShort(myBuffer, groupDB->name.length());
+	myBuffer = storeShort(myBuffer, (short)(groupDB->name.length()));
 	myBuffer = storeString(myBuffer, groupDB->name);
 
-	myBuffer = storeShort(myBuffer, groupDB->addresses.size());
+	myBuffer = storeShort(myBuffer, (short)(groupDB->addresses.size()));
 
 	for (unsigned int i = 0; i < groupDB->addresses.size(); ++i)
 	{
 		myBuffer = storeShort(myBuffer, 4);
 		myBuffer = storeInt(myBuffer, groupDB->addresses[i].id);
 	}
-	myBuffer = storeShort(myBuffer, groupDB->addresses.size());
+	myBuffer = storeShort(myBuffer, (short)(groupDB->addresses.size()));
 
 	for (unsigned int i = 0; i < groupDB->addresses.size(); ++i)
 	{
-		myBuffer = storeShort(myBuffer, groupDB->addresses[i].value.size());
+		myBuffer = storeShort(myBuffer, (short)(groupDB->addresses[i].value.size()));
 		myBuffer = storeString(myBuffer, groupDB->addresses[i].value);
 	}
 }
